Many Event Logic and Visual improvements
# Multiline event saving
Allows to load events from multiple lines. To do so, the loader asks the event if it thinks it is full and if not it adds the next line and asks again.

The text event will say it's not full if it ends with `\`
The shortcode events will say they are  not full if they don't end with `]`
The method in is `is_string_full_event(string) -> bool`
This means oc that you can also store events in multiple lines.

# Making things more modular
I have once again made a lot of the logic behind the events more modular.
This time it was about making the scripts not depend on specific events.

One exception is the DialogicEndBranchEvent!
This event is super special and I cannot see a way to make it deletable.
The Choice and Condition events however are now not needed anymore.

Instead events have a couple of more flags and methods they can fill in:

- can_contain_events :bool 
    -> marks events that open a new block and need an end event
- needs_parent_event : bool
     -> marks events that want to be the child of a specific event (like choice wants to be child of text)

These are mainly used for indentation logic (both in editor and when saving).
If one or both of these are turned on a couple of overridable methods can be used:
- is_expected_parent_event(event) -> bool
     -> to indicate what event a "needs_parent_event"-event needs
- get_end_branch_control() -> Control
     -> a control that will be added to the end branch

# New event lines
The editor has new lines connecting the events!
They also use the logic from above.

# Todo:
While the editor itself has been cleaned from these references, the event logic itself will need to be tackled next. This is a bit more complicated, but I will give it a shot. Main things I will be working on:
- Subsystem_Choices
- EndBranchEvent
- ConditionEvent
- ChoiceEvent