Add default portrait and mood (#1073)
Adds a default_portrait property to the character resource, an editor field in the character editor and an option to the portrait picker on the Join event (the new default)

Adds a default mood setting to the TypeSound editor. This mood will be used if no custom mood is set for the current portrait.