Add glossary
# Glossary resource
There is a new resource called DialogicGlossary.
It has two properties: 
- entries (Dictionary)
- enabled (bool)

# Glossary editor
There is a new settings editor for glossaries that allows to create new glossaries and edit their entries.

# Glossary subsystem
There is a new glossary subsystem. The most important part is the parse_glossary method that is used by the Text event.

# Glossary event
There is a glossary event, but it is currently empty. 

Other:
- moves the example scenes into an folder in the ExampleAssets.