Add pause() and resume() methods
This adds a pause() and resume() methods to the DialogicGameHandler.
These will propagate down to the Subsystems and set the paused boolean of the DialogicGameHandler.

Pausing is implmented for:
- text/input/voice
- sound/music
- choices (won't trigger when paused)

Making the character animations pause did not yet work, because the portraits themselves are a bit broken a the moment. Also I haven't yet found a good way to get the tweens to pause them.