Add next indicator node and functionality (#1130)
# Next Indicator Node
Adds a NextIndicator node. Basically a control node that will be shown and hidden. This allows to be very flexible with the design of the indicator. It has two properties:
-  show_on_questions
- show_on_autocontinue
Both are false by default.

# Functionality
The Text subsystem has two new methods to show and hide the indicators. I've added a 'text_finished' signal to the Text subsystem, that is now used by the Text event instead of the endless loop.