Separate text-to-events logic (#1129)
* Separate text-to-events logic

Mad a new TimelineUtil that holds the
- text_to_events
- events_to_text 
methods. These are now used by the TextEditor and the ResourceSaver/Loader.

The following code snippet works for me:
``` 
        var tmln = DialogicTimeline.new()
	tmln._events = TimelineUtil.text_to_events("""
	Jowan (Surprised): Wow this is interesting!
	- Yes
		VAR MyAutoload.exitement += 20
	- No
		VAR MyAutload.exitement -= 10""")
	Dialogic.start_timeline(tmln)```

* Fix saving 

The Text editor was saving even if it was hidden. That meant that when using the visual editor the text editor would always clear the file.

Also hides the VisualEditorToggle button when a character is loaded.

* Translation fixes

The translation settings where broken (buttons were not connected anymore).
Also the translations updates had a small problem but now it seems to work again.