2.0 Extra Data on Character Event (#1138) (#1160)
Allows to add a string of information on Join and Update:
`Join Emilio 3 [extra_data="Cool Sword"]`
If you use a custom scene, you can add a method called `func set_portrait_extra_data(data: String) -> void:`
If such a method exists, it will be called with the data as the argument.

This has no editor field for now. Showing it in an advanced mode was considered. It can be used in text-mode.

Authored-by: Danielle Cheney <thebardsrc@gmail.com>