Add a special case for text speed 0
In a text-speed of 0, godot's timer logic breaks down. A start(0) call does not set the time, causing unexxpected pauses. Bypassing that with a special case that bypasses the timer skips vital signal connections. This caused the event to stall and get stuck.

As a compromise, this patch forces the timer to run a short cycle at the start of the text reveal in order to allow the signals to connect correctly.

This only occours when the text speed is set to 0.