Text sped 0 special case (#1224)
* Add start and stop logic to CharacterVoice.gd

also fixed a naming error in dialog_node

* Add some lines of code I forgot to save before committing

it happens. And it is always a shame when it does.

* implement start_time and stop_time in audiopicker

The changes to the preview feature works as expected

* Add collapsable secund row to audiopicker

Known issue:
conflicts with EventPart_VoiceEditor.gd
The labels column does not change size to match the resizable audiopicker

* Merge Label and Voices containers into a grid container

This has solved the issue from the previus commit.
This is however a hacky fix, and may have unforseen consequences.

* Change the collapse icons

* First UI test

* Second UI improvments

For now the Audio-Region section will be disabled for Audio Event and Background Music event.

* Just added a tooltip

* New AudioPicker layout/design

Was trying to make it clearer what belongs together. Hope it works.

* Fix embarresing typo

I had a feeling something were wrong, now I fixed it.

* Updated Changelog

Apparently there wasn't even a note about the Voice Line System... ;-/

* In in a nw=v feature

With this, the text event will skip to the next, the moment the voiceacting is complete.
It is a hack, and needs to be better refined to fit in base code, but it works.

Tests work as expected.
When tested on a line without voice, it defaults to a 0 secund skip

* Add a special case for text speed 0

In a text-speed of 0, godot's timer logic breaks down. A start(0) call does not set the time, causing unexxpected pauses. Bypassing that with a special case that bypasses the timer skips vital signal connections. This caused the event to stall and get stuck.

As a compromise, this patch forces the timer to run a short cycle at the start of the text reveal in order to allow the signals to connect correctly.

This only occours when the text speed is set to 0.

Co-authored-by: KvaGram <0kvakk@gmailcom>
Co-authored-by: Jowan-Spooner <raban-loeffler@posteo.de>