Timeline Simplify saving, add emty_line preserving, fix label (#1347)
# Saving simplifications
The timeline saving logic has been simplified.
The TimelineResourceSaver now accepts both Timelines with text arrays as well as event arrays. If an event_array is present, it will convert it to text.

The visual timeline editor saves with a event array, while the text editor saves with a text array.

Note: I've removed the auto-dependency fixing for timelines. That is unfortunate, but necessary because it broke once loading didn't return a timeline with event resources. Maybe in the future there is a way to recreate this.

# Empty lines
The number of empty lines is now stored in the next event (in the events.empty_lines_above property). This allows to save them even when events are resources.

# Label fix
The name field of the label event wasn't working