Move global portrait settings to own tab and autosave
# Autosave on all changes for the character editor
This is not optimal, but is a quick fix for the portrait preview not working. Also character resources aren't complex enough to make this a problem I believe.

# Move global portrait settings to seperate tab
This makes default portrait and main_scale, *_offset and *_mirror part of the character editor again in the portrait section. This makes the editor less crowded.