Add export overrides section to custom portrait scene settings
New portrait settings section "Exports" if a custom portrait scene is selected. This tab shows all the exported variables of that scene and allows to set those.

The editor will now also keep the same scene if _should_do_portrait_update is defined and true. This allows blending from one portrait to the other in editor.