Add group and ordering support to portrait list
The groups are just a fake representations of portraits named with / in them. 

So "Emotions/Happy" will be represented as a group Emotions with an item Happy. 

As this was quite the change from the single level list we had before, many things like searching had to be redone. Search will only look for portrait items (not group items).