Test/simplify jump/return logic
Creates a new jump subsystem that holds all the jump related methods.

Makes the jump_stack part of the current_state_info so it's loaded and saved correctly.

Makes empty jump events jump to the beginning of the current timeline.
# 1318

Makes sure jumping to a label doesn't wait for another input. #1358