Character Editor Update (#1380)
* Move global portrait settings to own tab and autosave

# Autosave on all changes for the character editor
This is not optimal, but is a quick fix for the portrait preview not working. Also character resources aren't complex enough to make this a problem I believe.

# Move global portrait settings to seperate tab
This makes default portrait and main_scale, *_offset and *_mirror part of the character editor again in the portrait section. This makes the editor less crowded.

* Make sure portrait preview gets cleared if no portrait

* Add export overrides section to custom portrait scene settings 

New portrait settings section "Exports" if a custom portrait scene is selected. This tab shows all the exported variables of that scene and allows to set those.

The editor will now also keep the same scene if _should_do_portrait_update is defined and true. This allows blending from one portrait to the other in editor.

* Small fixes

- makes sure typing pitch cannot be set to <= 0 by randomness.

- forgot to stage the code that actually makes the export overrides work in game

* Add group and ordering support to portrait list

The groups are just a fake representations of portraits named with / in them. 

So "Emotions/Happy" will be represented as a group Emotions with an item Happy. 

As this was quite the change from the single level list we had before, many things like searching had to be redone. Search will only look for portrait items (not group items).

* Make sure non existing portrait doesn't cause problem in default scene

* Small fix