Translation improvements and fixes (#1396)
* Mainly Translation fixes

# Translation
- added syntax highlighting for the translation id (#id:...)
- added status message to indicate an action happening when pressing the Update and Collect buttons.
- fixed wrong reference to timeline processing method
- fixed saving failing saving of timelines (for newly added translation ids)
- hiding translation settings when translation isn't enabled.

# Other changes:
- Settings editor/s renamed to snake_case
- DialogicEditor._clear() method
Can be implemented to clear the state and content of an editor. This is implemented in the timeline editor, because when indexing new events for translation changes were getting lost on the currently opened timeline.
- fix in ComplexPicker for metadata not being set correctly when searching.

* Adding sections and testing locale setter

Added a picker for the testing locale. This is just getting/setting the internationalization/locale/test project setting, but makes it easier for users to set if they don't know the project settings to well.

Also added sections for readability.

* Add confirm dialog and message for removing translations

* Fix mistake

Forgot to stage the main scene which had a broken reference to the settings editor due to rename.

* Important bug fix

There was a wrong reference to a project setting. Because I had the translation_enabled setting from way back, I didn't notice.