Make creation of extensions (custom events, subsystems, etc.) easier
Adds a section in the general settings for creating a custom event &/ subsystem. This can be expanded upon in the future to also allow auto-creating settings pages, text effects and whatever else is possible to do in the future. 

Also makes sure custom additions are actually listed correctly by the plugin.

- DialogicEvent._get_icon() will now not return warning icon because it looks strange. Also no error if no icon exists. 
- General Settings page now uses FilePicker instead of custom labels+buttons
- default_dialog_scene setting moved from dialogic/editor to dialogic/.
- renamed dialogic/custom_events_folder to dialogic/extensions_folder