Replace portrait position by portrait container
This creates a portrait container that provides more flexibility in the portrait placement and scaling. This allows easier integration into UI's but also just in general more/easier control about the size of portraits/responsive placement.

Portraits are now added as children of their actual portrait containers instead of a canvas layers, so the PortraitHolder was removed.

# Other small changes
- make sure portrait data is correctly duplicated
- added change_portrait_mirror, change_portrait_extra_data, change_z_index methods to make change_portrait less crowded.
- fixed typos in character event field labels
- fixed typos in Position event
- added a preview character. It's the default preview character for the portrait_container
- added lot's of static tpying
- updated default layouts to use portrait container
- deleted broken BigRPG layout
- added helper method to character resource that allows to safely get a portraits info or alternatively the default portraits info