Simplify and fix History system  (#1461)
* Simplify History and fix bugs 

This update fixes all the previous issues with the history system such as the button not hiding, crashes on character leaves. 

It's also simplifiying the way the history system works.
There is now three types of history the history system can save:
- Simple history
    - events can log information here, that can be used to display a backlog
- Full Event history
    - stores a copy of all visited events in an array
- Already Read history
    - remembers what text events you've already read and emmits signals to use that info

Splitting these systems makes the code much cleaner and the individual goals easier to maintain.

* Small fixes

* Typo fix and AlreadReadIndicator removed

Removed the ugly test indicator for the already read history. I've left the script as an example.