Fix for skippable bug and custom layout setting (#1476)
* Fix missing default skippable value

* Make sure custom layout scene can be set 

There was some missing connections there. I also noticed that the layout scene path wasn't saved in the settings layout category so this path changed now. Small incompatibility with the previous version as a result.

* Make sure custom layout scene is used for testing

I've also removed the custom_testing_scene settings that was very redundant right now as it couldn't be set manually and was always the same as the layout_scene setting.