add null check to text update_name_label
There are a few places where we do a `update_name_label(null)`, which
makes `character.resource_path` throw an invalid index error. This
commit fixes this bug; also a fixes a related bug in the same function
where it is possible to access
`dialogic.current_state_info['character']` before it is defined.