Add autopausing-capability
Adds autopausing for certain letters. The sets of letters can be changed in the DialogText settings page.

The autopausing is done as a modifier that adds [pause] effects after these letters. A regex is used so letters inside of [brackets] or {brackets} are not altered. This is mainly important for the dot (.).