Add modules manager and first default override ability
There is a new settings tab that allows managing all the modules and extensions. It has a fully functional tree list of all the modules and their parts.

There is a editor for the event defaults. If changed these will be stored in a event default override dictionary. From there they are automatically set in each event before the actual event is loaded. Shortcode events also take these defaults into account for saving (won't add a value if it's the custom default).

I had to make sure these custom defaults are also applied to new events in the visual editor (as they are not acutally "loaded" there if new).

# Other
- fixed capitalization in Sound and Music event