Add default values for text event
You can now specify default values for the text event (character and portrait settings). 

For this to be fully functional, I had to add some syntax flavors. 

With no default character and portrait set:
`This is some text` will be said by noone
`John: Some text` will be said by John with his default portrait

With a default character set:
`This is some text` will be said by the default character
`_: This is some text` will be said by noone

If a default portrait is set for the event, it will have a higher rank then the default portrait of the character. To revert to the latter, you can write:
`John(): Some text` which will be said with Johns default portrait even if a default portrait is set for the text event.