Add modules manager and first default override ability (#1531)
* Add modules manager and first default override ability

There is a new settings tab that allows managing all the modules and extensions. It has a fully functional tree list of all the modules and their parts.

There is a editor for the event defaults. If changed these will be stored in a event default override dictionary. From there they are automatically set in each event before the actual event is loaded. Shortcode events also take these defaults into account for saving (won't add a value if it's the custom default).

I had to make sure these custom defaults are also applied to new events in the visual editor (as they are not acutally "loaded" there if new).

# Other
- fixed capitalization in Sound and Music event

* Complete edit fields actually working

Had only connected the line edits before for testing, now all kinds of edit fields should work.

* Add default values for text event

You can now specify default values for the text event (character and portrait settings). 

For this to be fully functional, I had to add some syntax flavors. 

With no default character and portrait set:
`This is some text` will be said by noone
`John: Some text` will be said by John with his default portrait

With a default character set:
`This is some text` will be said by the default character
`_: This is some text` will be said by noone

If a default portrait is set for the event, it will have a higher rank then the default portrait of the character. To revert to the latter, you can write:
`John(): Some text` which will be said with Johns default portrait even if a default portrait is set for the text event.

* Make character settings defaults overridable

Makes most of the character settings overridable. 
Animation name, length and wait are left out to avoid a crash with the already existing "ActionType"-dependant default settings in the Portrait settings tab.

Also added a scroll bar to the modules manager page.

* `text_submitted` -> `text_changed` to avoid having to press `Enter`

---------

Co-authored-by: Emi <2206700+coppolaemilio@users.noreply.github.com>