Make sure variables aren't always reset
For most games you want variables to be persistent unless loading a new save.
However many things should be reset when a new dialog opens. So I've added clear-flags to the dialogic game handler. These can be used with the clear() method. Most of the time now (at timeline end or dialog start) it won't reset the variables (only at game startup or loading a save).

This system should be expandable.