Add animation support (#1547)
* Add safety check when interrupt-loading a timeline

When opening a visual timeline while another one is still loading (fairly common with longer timelines) the last batch would always still be added to the newly opened timeline. A new check now waits for the last batch of the previous timeline before clearing the timeline.

* Initial tests for animations

* Better new text animation

* Make sure variables aren't always reset

For most games you want variables to be persistent unless loading a new save.
However many things should be reset when a new dialog opens. So I've added clear-flags to the dialogic game handler. These can be used with the clear() method. Most of the time now (at timeline end or dialog start) it won't reset the variables (only at game startup or loading a save).

This system should be expandable.

* Fix merging issues

* Remove print