Fixes for save/load systems (#1561)
* Fixes for saving/loading

Fixes for #1560 as described there

* Save/Load fixes & improvements

Makes sure the name label reappears when loading.

Also splits the Dialogic.start() functionality into to parts (which are both still done by start(). This allows adding the layout node without starting a timeline. This is useful before a load() call, to make sure the game can be displayed.