Improve text systems and TextBubble layout system
# Text systems:
- adds `about_to_show` signal to text subsystem. 
- adds enabled flag to text node

# TextBubble Layout
The text bubble layout will now display a different text bubble for each "registered" character and use a fallback one (that has no position tracking) for all other characters.
Characters can be registered with the register_character() method that takes a dialogic character resource and a node2D to track to.