Silence error on get_meta(..., null)
As per the `Object` docs, providing a default value of `null` to the
`get_meta` call doesn't silence the error in the event that the metadata
key doesn't exist. We have a few instances of this pattern throughout
the codebase, which were producing unintended error messages under
certain circumstances.

This commit replaces them with explicit `has_meta` checks,
where appropriate. In situations where the null value should be an error
(typically because it was previously checked) I removed the default
argument, to avoid confusion (the behavior is the same).

Also did some boy scout rule cleanup of trailing whitespace and
refactored some confusing branch logic in `DialogicGameHandler.gd`.