Many fixes (#1701)
* Fix settings event RESET ALL to text

* Change variable event syntax to `set` instead of `VAR`

* Add \ logic into text event

The text event will now automatically add a \ if it would be mistaken for another event otherwise. Any slash at the very beginning will automatically be removed when loading the event.

* Make sure deleted events aren't referenced as selected anymore

* Add name_label_root setting

This makes the hideWithScript script obsolete. Works the same as on the dialog text node now.

* Add support for nice names for enums

If you provide a suggestions function for enums, you can now also specify text_alternatives which will be favored by the loading/saving logic.

* Fix indentation/line drawing when adding new events

* Update lines when expanding event body