Events: New call event (#1858)
Adds a new event that replaces Call Node. 

The new event has the following text syntax:

      do Autoload.method(arg1, arg2, arg3)

The new visual event has pickers for autoload and method. The arguments can now be of complex types. There is warnings if the number of arguments is wrong or an argument type is wrong.

This new event ALWAYS awaits.
Inline features have been dropped in favor of custom text effects.