Improve unique-identifier-directory handling (#1950)
This moves the unique-identifier directories to a new DialogicResourceUtil class.
It simplifies the creation, access and modification of these directories.

Importantly the timeline and character directory are now saved in the project settings. This allows us to make them editable.

This also adds a Unique Identifiers tab to the Reference Manager window where you can edit them.

This PR does a bunch of other restructuring as well.