Improve custom portrait workflow
This makes it so a custom layout will auto-reload when it's script was changed. Unfortunately it's pretty hard to listen to a scene getting saved right now, but that will be implemented in the future, once it's easily doable.

Also makes sure the simple_highlight_portrait doesn't dim in editor.