Added Vec3/4 and made some improvements to Number element
Added Vector3 and Vector4 types, unified under a common base, so people making custom nodes have a unified way to write that data. Often people will want to use nodes that manipulate 3D space data, like the Camera or 3D character positions for example, so Vector3 is important.

Updated some of the code for the Number field as well to support more display customization, partly to clean up Vector display.  The increment/decrement button can be optionally disabled through an extras flag, as well as a left side label tied to the number field itself.

Split off the icon resource into a dedicated resource that can be shared among elements, and also split it in half so up and down can be visually distinct and clear on which you're clicking.