Fix a loading issue, improve logic
Didn't realize len() didn't work on the Vectors, but the section I used it is rarely called so didn't notice.

Refined the Vector classes to organize the logic flow better.

Got rid of a bug with the Number class that was resulting in max precision long numbers showing up in the fields upon restart despite "enforce_snap" being on.