Vector types and number improvements (#2094)
* Added Vec3/4 and made some improvements to Number element

Added Vector3 and Vector4 types, unified under a common base, so people making custom nodes have a unified way to write that data. Often people will want to use nodes that manipulate 3D space data, like the Camera or 3D character positions for example, so Vector3 is important.

Updated some of the code for the Number field as well to support more display customization, partly to clean up Vector display.  The increment/decrement button can be optionally disabled through an extras flag, as well as a left side label tied to the number field itself.

Split off the icon resource into a dedicated resource that can be shared among elements, and also split it in half so up and down can be visually distinct and clear on which you're clicking.

* Fix excessive spacing between Vectors

* Remove Value spinner styling changes

* Revert "Remove Value spinner styling changes"

This reverts commit 3a4d94a72e3c1e2848c5a8979fae147593d8ff84.

* Various improvements to Number input UI

-Added back vertically split value change arrows with independent events and highlighting
-Value box and arrows respond to arrow keys to adjust value accordingly
-Key or mouse can be held to start a continuous adjustment of value
-Suffix and Affix added as integrated Labels to the side of input field, keeping things tied together while not being a part of the actual value.
-Adjustment arrows also encapsulated by the input field in the same manner.
-Adjustment arrows will disable themselves when hitting a Min or Max value to visually indicate a limit was hit.

* Fix a loading issue, improve logic

Didn't realize len() didn't work on the Vectors, but the section I used it is rarely called so didn't notice.

Refined the Vector classes to organize the logic flow better.

Got rid of a bug with the Number class that was resulting in max precision long numbers showing up in the fields upon restart despite "enforce_snap" being on.

* Various refinements and fixes

- Fixed number fields causing line they are on to expand extra vertically
- Fixed number fields pushing elements to the right of them to the far end
- Touched up some code styling

* Fix auto-increment trigger when rapidly clicking, style improvements

The small delay before activation of the auto-incrementing allowed for situations where click was released during the delay, but pressed down again before the delay was finished, resulting in auto-increment triggering when it shouldn't, resulting in inconsistent number changing behavior.

* Get rid of dedicated style assets for number field arrows

Not a style that was going to be reasonably used by other things, so integrated it into the scene file to reduce file clutter

* Improve number styling/behaviour

- on focus select all
- extend field box to full number field
- pad decimal numbers based on step

* Small fixes to scripts and new inclusion of vector prefixes

---------

Co-authored-by: Jowan-Spooner <raban-loeffler@posteo.de>