Audio event bus management simplification (#2212)
The Music and Sound events will now default to "" as their bus, and when given an empty bus won't change the bus of the default_music_player or default_sound_player.

This allows code like this, to change the default bus, without changing the event default:
`Dialogic.Audio.base_music_player.bus = "Music"`
Same for sounds.

Previously it would always default to the "Master" bus, except if you changed the event default. Assigning a default to the event in the modules manager still overwrites this. And both default audio players also still default to Master, so this does not break compat.