Improve choice logic (#2252)
This separates the choice showing into three parts (that can be used somewhat independantly of each other):
- Collecting question information (choice subsystem)
- Showing choices (choice subsystem)
- Loading choice info (choice node!)

This also adds logic to store visited choices in the already_visited history, and adds info on whether a choice has been visited to the choice_info dictionary. This should make it pretty easy to change the style of choices, when they have already been visited.