Choice event enhancements (#2304)
This changes the syntax of the choice event a bit:
```
- Choice Text | [if Condition] [shortcode="value"]
- Choice Text | [shortcode="value"]
- Choice Text
```
*The old syntax without | still works though.*

This also adds a new extra_data field to choices. This data will be provided to choice buttons so they can now be uniquely identified and given arbitrary information. This would be even cooler, if the dictionary field would be nicer looking. 
- fixes #2286

This also makes it so you have to press a button to enable the condition on choices, which looks much better. 
- fixes the UX problem #2272

This also fixes a bug with visual editor fields not having the correct value if they start hidden and are then revealed by a condition changing.