Improve animation naming/identifying logic (#2293)
This is a mix of a general improvement of the SpecialResource system and the PortraitAnimations in particular. 

Portrait animations now specify different versions of themselves with different names.

This system could potentially be extended so that portrait animations can expose arguments, but that is a challenge from the event syntax. I'll have to see whether this is feasable.

This should eventually fix all regressions of animations being seemingly interpreted/named wrong.