Improve custom portrait UX (#2315)
Adds a custom portrait scene browser that shows available portrait scenes and portrait scene Presets.

There are four types of entries:
"DEFAULT" will simply set the portrait scene to "" (empty string), and thus default to the scene set in the settings.
"CUSTOM" allows selecting any scene from the file system.
A "GENERAL" item will use the specified scene.
A "PRESET" item will create a copy of the scene and then use that copy. This makes it easy to create certain scenes from a base preset. 

This moves the Layered Portrait and the SimpleHighlightPortrait into individual modules for a bit better organization.

Also:
- Some code cleanup
- Changes the layered_portrait scene to be a cleaner Preset
- Adds a "fix offset" setting to the layered portrait script that attempts to correct the offset of the portrait based on the first node, there is probably situations where you wouldn't want this, so it can be turned off, but it makes the creation a bit more fool-proof.