Add option to include simple and/or full event history in savegame (#2320)
* History: Add option to include simple and/or full history in savegame
Simple and full event history can be toggled separately. If one option is enabled:
- Include that history in Dialogic.get_full_state() / Save.save().
- On Dialogic.load_full_state() / Save.load() replace that history with the history from the savegame.
  - If the savegame doesn't contain that history, clear it.
  - Events appended to the full event history between the clear_game_state() and the load_game_state() are dropped intentionally.
- On Dialogic.clear(FULL_CLEAR), clear that history.

The changes to the text event are needed to avoid duplicating the last entry in the simple history after a save+load. This also makes the text show up immediately after loading instead of the normal typing style.

* Code Style Fixes

Adjust some code to follow the gdscript style guide.

---------

Co-authored-by: Jowan-Spooner <raban-loeffler@posteo.de>