Cleanup portrait animations and fix speaker portrait slide-off
This renames DialogicAnimation `end_position` and `orig_position` (which where the same) to `base_position` and also adds `base_scale`. A slightly breaking change for people who made their own custom portrait animations.

It also adds an `await.process_frame` on change_speaker() when joining a character as it can happen that this is called before the textbox is shown and thus before the portrait container is correctly sized.

Disables the non-functional signal `portrait_animating` for now until I find the time to make a proper re-implementation.