Portraits: Fix portrait sizing when window content is being stretched/scaled (#2445)
Closes: #2443

Based on conversation with zaknafean, I checked how this was done in Dialogic 1, and it uses get_viewport().get_visible_rect(), which according to my testing does respect the window stretch mode in Project Settings.

More detailed testing may be required, I only tested that the size matches expectations based on the stretch mode settings of disabled and canvas_items. When set to disabled the size reflects the actual window size, but on canvas_items the viewport size is returned instead, regardless of the window size. If I have time for further testing I will post any results as well.