Styles: Rework internal structure of styles (#2402)
This is a big rewrite of how styles are stored in layers with the main goal of identifying styles not by their index but by an identifier. This solves a big issue that inherited styles previously had, where moving the layers of the base style was allowed but would break sub-styles.

This also includes some code which auto-updates existing styles, but I'd like to remove this code again before we enter the beta phase (or at least for stable).