Visual Editor: Fix for file field (#2473)
The file event field would cause the path to get changed to just the file name without the directory when you accidentally click on the field and click off the field without changing anything. This changes adds a check to make sure the current value and field text value aren't the same before calling _on_file_dialog_selected, or, if file_mode isn't directory open mode, that the current value as just the filename isn't the same as the field's value.