Improve Visual Timeline Loading and Search Logic
Adds a loading Bar to the visual timeline editor, which is useful for long timelines, so you can more easily tell how long loading will take.

Also fixes some issues with the searching logic in the visual editor, which made it look like it crashed.

Also makes it so changes made during the loading of a long timeline will be discarded in favor of removing everything not yet loaded from the timeline if the editor is switched before the loading finishes (previous behaviour).