Feature: Play from here (#2535)
* Implement play from here

Adds a play from here option to the right click menu both in the visual editor and the text editor. The action can also be triggered with Ctrl+F6 or Ctrl+Shift+B on macOS. Those shortcuts have been added to the shortcut cheat sheet.

The text editor thing required a new method on the timeline resource that I'm not super pleased with, but it works. It works by keeping an index of all the lines and the event index they end at. I hope this doesn't have a big impact on timeline processing, but I don't think it should.

* Fix text event doc link

---------

Co-authored-by:  Pasquale J Crea <pasquale@firedfromlife.com>