Improve runtime resources (timelines and characters created through code) (#2537)
* Fix for dictionary field in custom portrait export overrides

* Allow registering identifiers for runtime resources

This makes both DialogicTimeline and DialogicCharacter extend DialogicIdentifiableResource which basically adds a get_identifier and set_identifier method. If an identifier is set at runtime for a non-saved resource, the resource is registered to the directories directly (without a path) allowing for the use of those resources in timelines from then on.

This also once again allows using "invalid" character names in timelines and will now even interpret the portrait as a color, which allows to have one-off characters say something. However the portrait picker is not shown in the visual editor for those characters for now so this is kinda a text-editor exclusive rn.