Improving code/fixing issues with update_speaker and portrait highlights
This code is defenitely breaking some compatibilty, because the speaker is now stored as an identifier and not as a resource_path in the current_state (breaks compat with old saves and maybe with some custom code of people).

Could also break stuff because the order of operations in the text event has slightly been changed.

Also removed the unnecessary get_current_character from the Portraits subsystem, as it just did the same as the Text method, and speaker is mostly a text concept anyways.

Fixed a dumb mistake in the dialogic_identifiable_resource