Portraits: Improving/fixing issues with Speaker signal and Highlight methods (#2543)
- fixes several issues reported on the discord server

This code is defenitely breaking some compatibilty, because the speaker is now stored as an identifier and not as a resource_path in the current_state (breaks compat with old saves and maybe with some custom code of people). This is necessary to allow the speaker_update method to work with runtime-characters.

Could also break stuff because the order of operations in the text event has slightly been changed.

Also removed the unnecessary get_current_character method from the Portraits subsystem, as it just did the same as the Text systems method, and speaker is mostly a text concept anyways.

Fixed a dumb mistake in the dialogic_identifiable_resource