Improve slide animations
This changes the slide animations. Slide In/Out Left and Slide In/Out Right are now renamed to Slide From Left, Slide To Left, Slide From Right and Slide To Right. All the slide animations have been adjusted to slide exactly the needed amount so that the portrait fully leaves the screen, not just sliding a fixed amount. To do this, the animation class has gotten two new methods that help determening the animated nodes (e.g. Portraits) size and position.

This breaks compatibility because those animation names have changed.