Improve resource dependencies & Add resource icons (#2637)
Implements/Fixes the _get_dependencies methods for timeline and character resources. This means you can now right click a timeline or character resource in the file system and use the "Edit Dependencies" option to see all the resources used by that file. For characters even the editing/replacing works.
To make this work in events, in the get_shortcode_paramaters() external references can now be marked. 
I'm not sure if the timeline stuff has too big of a performance cost, as it needs to open the file to list the dependencies and I think godot does this whenever the file is saved. If it turns out this slows down big projects too much, we will have to reconsider this.
- somwhat adresses #2605 

This also adds a new timeline icon (instead of using godots three-stripes icon) and sets the resource icon for characters and timelines (so they finally appear in the file system!!)
- adresses #2355