Fix issues with runtime/variable character colors (#2660)
* Fix issues with runtime/variable character colors

- Fixes #2652.
You can now use expressions to determine the speaker again, without it automatically creating a new character. The expression will be parsed in the text events "_execute" method and if it results in a valid character identifier that character will be used. Otherwise a runtime character will be created as expected.

* Make sure text events trailing empty newlines don't make a mess

- fixes #2608