Change Saving to use DialogicSaveState resource
Introduces a new DialogicSaveState resource. This is the first step in many to make the saving logic more modular and completely remove Dialogic.current_state_info.
For easier debugging, the save system saves this as a .tres for now, which makes it impossible for encryption to work, this can be changed back later.

Subsystems now have the pack_state() and unpack_state() methods, which collect it's exported variables into a dictionary.