Finish textbox visibility cleanup & Improve subsystem docs
Allows subsystems to store extra_state on save and use it on load. This is mainly intended for state outside the subsystem script, e.g. on nodes in the style or custom parts like portraits, backgrounds, etc.

The textbox visibility is now handled per "textbox-identifier", which can be specified on the dialog_text node. A new "active_textbox" can be changed via script.