Change clear event to use textbox fade animation length
This slightly changes the behaviour of the clear event. Instead of the given time, it will now just await the textbox hide animation.

This is good because previously it would default to a second of delay when the textbox animation was instant, which felt very unexpected to users, especially now that a clear event is automatically added at the end of dialog.