Add a setting to events to prevent them from expanding on creation
Though on most events with a body it's useful to have them open when first created, on some it's kinda annoying imo. So for now the label, choice, wait input will stay collapsed when created.
My reasoning is a bit different for them. For label and wait input I think you mostly don't need those settings. But for choices it's mostly because you will often add a bunch of choices at once and then having them all expanded is unnecessary.